Correcting Object RAM layout
The $200 bytes of RAM previously denoted as Object RAM does not seem to ever be used. There are times where it looks like everything from Object_RAM to Object_RAM_End should be used, but only Object_RAM to LevelOnly_Object_RAM_End actually is. Because of this, I think it's clear that Object_RAM_End is in the wrong place, and that $200 bytes of RAM is completely unrelated.