Added another sound driver bugfix
If you play the final boss music, and then play the extra life jingle,
the trumpet-sounding FM instruments will play really loud before fading
back in properly.

This is caused by an overflow: that instrument's TL is $60, and fading
adds $28 to it. The TL is a 7-bit value, so it wraps to $08, near-full
volume.