Converted build tools to C
Most of the code was already C, so this didn't need too many changes.
s2p2bin's FW-KENSC is very C++-centric though, so I swapped it out for
my Saxman compressor. Yay for smaller executables.

checkhash was already C, so this makes it so you only need one compiler
for everything. I also took the chance to make the Win32 Makefile and
build.sh use the same compiler flags.