Add accurate Saxman compression to s2p2bin
See this thread:
https://forums.sonicretro.org/index.php?threads/i-found-the-saxman-compressor-source-code.39261/

Basically, the source code to Sonic Team's original Saxman compressor
was found, so I've integrated it into s2p2bin in order to
accurately-compress the sound engine.

Oddly, it still doesn't explain why there's a '4E' byte stuck to the
end of the compressed data. It crashes the original C decompressor,
but the 68k version seems to just ignore it.