Sound driver documentation updates and stuff A more liberal take on things: some names have been replaced or reverted, based on feedback from my last commit.
Other things: I completely redid one comment for clarity's sake (since there was misleading information)... and I noticed this last time, but only saw fit to mention it now: there's a line that was changed between 1 and 2 that I don't know if this is a bug. Perhaps someone more knowledgable can clue me in?