Remove an unnecessary rev02even
rev02evens should only occur after 'dc.b' directives.

By the way, I'm pretty sure that the REV02 padding and instruction
optimisation differences were caused by the assembler being
different, be it a different version, configuration, or a different
assembler entirely.

Notably, asm68k has an option for automatically padding dc.w and dc.l
directives, but it works completely differently to how these dc.b
were padded. Additionally, asm68k also has options for automatically
optimising to PC relative addressing, and optimising additions and
subtracts to their quick forms, so it's not unheard of for assemblers
of the time to offer these features.

Because of the lack of 'dc.b' padding however, and the fact that the
source code in the Nick Arcade prototype makes use of an 'addsym'
directive as well as mid-line comments starting with a '*' symbol,
I think it can be concluded that Sonic 2 was not assembled with
asm68k.