Make the shared special stage RAM safer
Now there are proper checks to ensure that shared data occupies space
that it is meant to. For example, it was previously possible for the
shared special stage RAM to overwrite the DMA queue variables, with
catastrophic results. Now, if that happens, you'll get a build-time
error warning you of it.