Sod it: add a 'fixBugs' option
This disassembly exists to document Sonic 2, right? Well what better
way to document its bugs then to show exactly how to fix them?

Other reverse-engineering projects like the Super Mario 64
decompilation do this, so it's not exactly unheard of.

Still, we should be mindful to not include fixes for things that
aren't even bugs. I know that determining this can be extremely
tricky, as is the case with Sonic 1's spike bug.

If it ever gets to the point where people are seriously debating
whether a bug is a bug or not, I suppose there can always be another
option for 'debatable' bugs. Or maybe we should just avoid fixing
debatable bugs. I don't know.