Remove skipChecksumCheck from allOptimizations
'allOptimizations' is more intended for enabling psuedo assembler
optimisations, like turning redundant '0(aX)' operands to '(aX)',
and inlining JmpTos: anything in terms of optimising the actual
game logic should be an exercise for the user.

I've also removed the description of the checksum check as
'unnecessary', since it does have a proper purpose in ensuring that
the cartridge can be read properly. This is useful for running
software on a flashcart, since I've personally encountered instances
where the game didn't flash correctly, and the checksum check was
able to detect it and prevent me from being confused by strange bugs
and crashes. Calling it 'unnecessary' makes it seem like it's totally
pointless.

Personally, I think that 'skipChecksumCheck' and 'useFullWaterTables'
should be removed, as they don't document anything or exploit
pre-existing optimisation related logic: they're just changes that a
ROM-hacker might want to make.