Point out bugs in some assets
The 'fixBugs' flag can't fix these unless 'corrected' files are
included with the disassembly, which might be out of scope for this
project.

...Maybe this flag was a bad idea. Explaining bugs and mistakes is
good and all, but being the arbiter of which bugs should be fixed and
exactly *how* they should be fixed may be a step too far. Dammit.