Convert the build system to Lua.
See here:
https://github.com/sonicretro/s1disasm/commit/e5651cf7f3fd45934ba5e7926874de0c3bffcbde

Doing this effectively unifies the two build systems (Batch and Bash)
into one script that works everywhere, while not sacrificing the
usability of the disassembly for Windows users.

This change also brings this disassembly in line Sonic 1's
disassembly. We might even be able to unify the build systems across
the various disassemblies someday.