Use pre-built executables for all platforms.
I can't find AS nor Flamewing's modified AS as standard packages of
Debian or Arch Linux. I don't think the average user wants to go
through the process of compiling AS on their own and messily
installing it to their system just to assemble Sonic 2, so I'm
bringing this disassembly in line with the Sonic 1 and Sonic 3 &
Knuckles disassemblies by providing AS binaries as part of the repo.

While I was at it, I added binaries for s2p2bin and fixpointer as
well. Maybe I'll bring back the ability to automatically compile them
later. Though, if anything, I'd like to see s2p2bin and fixpointer
migrated to Lua so that binaries aren't needed at all. I worry that
a Lua-based Saxman compressor may be too slow though.

Unfortunately, this commit got tangled with a refactor of main Lua
build script, so picking apart the changes may be difficult.

Currently, the only binaries are for Windows and 64-bit Linux: Mac
and 32-bit Linux binaries will need to be added in the future.