Convert music and SFX to SMPS2ASM.
This wasn't as bad as I was expecting: Lua really is great for build
scripts.

As you would expect, the music is assembled and compressed at
build-time. Which songs are compressed and which are not is
determined by 'sound/music/compressed/list of compressed songs.txt'.

Also, mixed in with this big bastard of a commit is that I've added
a CMake build script for the build tools. This allows the Win32 build
tools to be built with MSVC. I had to edit s2p2bin.c to remove a
Variable-Length Array, but other than that everything built on the
first try.