Lua: Allow extra periods in input filename.
If the input filename was something like 'sonic.bum.asm', then AS
will truncate to the first period and output files such as
'sonic.log' and 'sonic.p'. The Lua script wasn't accounting for that,
and checking for 'sonic.bum.log' and 'sonic.bum.p'. This fixes that.

Thanks to Lil-G for pointing this issue out in #39.