Remove a bogus bugfix.
I thought that this code ran at the start of the special stage, but
it doesn't: it runs at the start of the special stage's results.

I swear I've seen a lot of corruption in the special stage's results
lately, and I bet it's this dumb "bugfix" that's been causing it.