Preliminary code split.
This split has been done in a way that tries to replicate the
original source. This is done by splitting whenever linker data is
encountered. Unfortunately, there are a few strange places where
there SHOULD be a split, but no linker data exists. This either means
that the original source code really was just messy, or that the
linker would not always emit data between files.