Overhaul JmpTo handling.
Now JmpTo labels are turned into their base labels through an AS
function. Because of this, it is no longer necessary to have both
labels in a 'jsrto' or 'jmpto' invocation.

I've also added a 'jmpTos' macro for automatically managing the
JmpTos themselves: this creates the labels, 'jmp' instructions, pads
the start with a 'nop' instruction or 00 bytes, and pads the end to
a multiple of 4.

Unfortunately, I cannot figure out how the original assembler decided
to pad with a 'nop' instruction or 00 bytes, so that is currently
hardcoded per JmpTo block.