Lua: Use better fallback values for unknown platforms.
This pattern matching was a dumb idea in the first place, but until
we can verify what the 'native' OS and Arch strings of each platform
are, it will have to remain for backwards compatibility.

Anyway, if a platform does not match any of the patterns, its strings
are set to 'unknown', which is utterly useless. So, instead, the
native strings are provided instead.